Address code review: fix int64 overflow, clarify magic numbers, replace en-dash in shell script
Agent-Logs-Url: https://github.com/xuanlocc2/crate-lm/sessions/85c46111-910f-466b-be2c-405a27c1ce7c

+++ data/wikitext/prepare.py
@@ -39,7 +39,7 @@ if __name__ == '__main__':
     )
 
     for split, dset in tokenized.items():
-        arr_len = int(np.sum(dset['len']))
+        arr_len = int(np.sum(dset['len'], dtype=np.int64))
         if arr_len == 0:
             print(f"Skipping empty split: {split}")
             continue
